Extract proposal contracts
