+++ packages/contracts/source/contracts/crypto/messages.ts
@@ -1,7 +1,13 @@
 import type { Enums } from "@mainsail/constants";
 
-import type { Block, BlockHeader } from "./block.js";
 import type { KeyPair } from "./identities.js";
+import type {
+	Proposal,
+	ProposalData,
+	ProposedBlockSerializable,
+	ProposedData,
+	SerializableProposalData,
+} from "./proposal.js";
 import type { AggregatedSignature } from "./signatures.js";
 
 export type MessageType = Enums.Crypto.MessageType;
@@ -18,37 +24,6 @@ export type WithoutBlockHash<T> = Omit<T, "blockHash">;
 export type WithOptionalBlockHash<T extends HasBlockHash> = WithoutBlockHash<T> & Partial<Pick<T, "blockHash">>;
 export type SignaturePrevoteData = WithOptionalBlockHash<SignatureMessageData>;
 export type SignaturePrecommitData = WithOptionalBlockHash<SignatureMessageData>;
-
-export interface ProposalData {
-	readonly blockHeader: BlockHeader;
-	readonly lockProof?: AggregatedSignature;
-	readonly round: number;
-	readonly data: { serialized: string };
-	readonly validatorIndex: number;
-	readonly validRound?: number;
-	readonly signature: string;
-}
-
-export interface SerializableProposalData {
-	readonly round: number;
-	readonly validRound?: number;
-	readonly data: { serialized: string };
-	readonly validatorIndex: number;
-	readonly signature?: string;
-}
-
-export interface Proposal extends Omit<ProposalData, "data"> {
-	isDataDeserialized: boolean;
-
-	readonly serialized: Buffer;
-
-	deserializeData(): Promise<void>;
-	getData(): ProposedData;
-
-	toSerializableData(): SerializableProposalData;
-	toData(): ProposalData;
-	toString(): string;
-}
 
 export interface PrevoteData {
 	readonly type: MessageType;
@@ -81,14 +56,6 @@ export interface Precommit extends PrecommitData {
 	toSignatureData(): SignaturePrecommitData;
 	toString(): string;
 }
-
-export interface ProposedData {
-	readonly block: Block;
-	readonly lockProof?: AggregatedSignature;
-	readonly serialized: string;
-}
-
-export type ProposedBlockSerializable = Omit<ProposedData, "serialized">;
 
 export interface SerializeProposalOptions {
 	includeSignature?: boolean;
